api_owl_vit/requirements.txt
flask
torch
transformers
Pillow
requests